[front/components/app] - refactor: rename Deploy to ViewAppAPIModal
 - Moved `ViewAppAPIModal` to its own file for better modularity
 - Updated imports in `CopyRun.tsx` and `[aId]/index.tsx` to reflect the new file location and name

[front/pages/w/[wId]/spaces/[spaceId]/apps] - refactor: update import path for ViewAppAPIModal

 - Adjusted the import path after the `ViewAppAPIModal` component was renamed and relocated
